Balance: Introduce external buzzer for footswitch faults
Some VESCs have unused pins which we're leveraging to control a simple
digital ON/OFF buzzer (or could be an LED too)

Trigger buzzer when switch_state is OFF (i.e. foot sensor fault) when above
min erpm. Stop buzzer when switch is back ON, or when timer expires and
a fault is triggered.

Signed-off-by: Dado Mista <dadomista@gmail.com>