Balance: Use buzzer to alert user about full-state switch faults at speed
In an effort to prevent unexpected shutdowns due to faulty foot sensors
we use the buzzer to alert the user when the switch turns off.

Must be used in combination with a longer full_fault delay.

This gives users a chance to react, either adjust their stance or
brace for impact or slow down...

The downside of a long fault delay is that the board keeps going if you
jump off - which is a common user behavior at low speeds.

In order to avoid this problem at low speeds, we want a reduced delay,
so we use the half_fault delay at low speeds.

In other words:
If I jump off at 3mph, don't keep the vehicle running for a long time.
If my footswitch fails at 15mph give me an opportunity for the buzzer
to alert me.

Signed-off-by: Dado Mista <dadomista@gmail.com>